ovsdb: Abort schema conversion before election timer expires.
Since the topic of clustered database schema conversion issues was
first brought up [0], there has been several performance
improvements which has greatly improved the situation.

However, we cannot get away from the fact that the process remains
unpredictable, when faced with a cluster that has grown to a scale
no longer supported by its configuration.

The OVSDB Server is mostly single threaded, and for clustered
databases the schema conversion is done server side in the main
thread.  While conversion is ongoing the server will not service
critical cluster communication, such as Raft elections.

In an ideal world operators would regularly re-assess their
deployment configurations with real data to ensure parameters such
as the OVSDB Raft election-timer is appropriately configured for
all cluster operations, including schema conversions.

The reality is that this does not happen, and operators are instead
discovering the hard facts of managing a OVSDB cluster at critical
points in time such as during database upgrades.

To alleviate this situation we introduce time keeping in the
ovsdb_convert functions for clustered databases.

A timeout is deduced from the configured election timer using a
divisor of 4 to compute buffer time.  Buffer time is the time
required for writing the result of the conversion to the raft log
and initiating command to replace data on peers before an election
is held for the next term.

For conversions initiated by a client (such as ovsdb-client),
the process will be aborted and an error will be returned if the
schema conversion process overruns the timeout.  Conversions
initiated from storage will not abort, a warning indicating the election
timer being set too low will be logged instead.

This will allow a human or automatic operator to take necessary
action, such as increasing the election timer, before attempting
the schema conversion again.

0: https://mail.openvswitch.org/pipermail/ovs-discuss/2022-December/052140.html
Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>