python: Allow ovs-flowviz to run without graphviz.
The graphviz library has a large set of dependencies, and as such
it may be desireable to not install it by default.

Before this patch execution of ovs-flowviz would end in
ImportError when graphviz is unavailable.

Exit with a friendly message instead, allowing the user to consume
the parts of the program that do not depend on graphviz.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>